requirements.txt
async-timeout==3.0.1
attrs==19.3.0
CacheControl==0.12.6
certifi==2020.4.5.1
chardet==3.0.4
discord.py==2.1.0
dnspython==1.16.0
idna==2.9
jikanpy==3.4.2
lockfile==0.12.2
msgpack==1.0.0
multidict==4.7.5
mysql-connector-python==8.0.19
protobuf==3.6.1
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
